+++ Wake On Lan/Routes/Routes+PostLaunch.swift
@@ -1,0 +1,24 @@
+//
+//  Routes+PostLaunch.swift
+//  Wake on LAN
+//
+//  Created by Dmitry Stavitsky on 11.11.2022.
+//  Copyright © 2022 Владислав Лисянский. All rights reserved.
+//
+
+import HostList
+import SharedRouter
+
+extension WOLRouter {
+    /// Navigates to `HostList` screen.
+    public func hostList() -> Route {
+        Route {
+            try? defaultRouter.navigate(
+                to: defaultStepRoutePushAsRootAction(factory: HostListFactory(router: self)),
+                with: nil,
+                animated: false,
+                completion: $0
+            )
+        }
+    }
+}
